refactor: extract metadata row methods to shared file
